Added https support
You need to define ASYNC_TCP_SSL_ENABLED as 1 to use it, which is what you will need to do for the whole project anyway. You will also need v2.4.0 of ESP8266 arduino and the git version of ESPAsyncTCP.

This is only partially tested, as it runs out of memory performing the handshake in my application!